refactor(examples): simplify shadcn Base UI forms
